+++ microservices/Work/controllers/works/worksData.js
@@ -3,13 +3,19 @@ const AdminSchema = require('../../dataBase/models/adminSchema');
 const moment = require('moment');
 const {client} = require('../../dataBase/Elasticsearch/connect');
 const Lookup = require('../../dataBase/models/lookup');
+const {Schema} = require('mongoose');
 
 module.exports.create = async (req, res) => {
     try{
         let {name, style, adminId} = req.body;
         if(!name || !style || !adminId) throw new Error('Some field is Empty!!');
-        const isPresent = await AdminSchema.findById({_id: adminId});
+        let isStylePresent = await Lookup.findOne({key: style});
+        if(!isStylePresent) throw new Error('Style NOT VALID!!!');
+        if(isStylePresent.type !== 'adminWork') throw new Error('Style NOT VALID!!!');
+        let isPresent = await AdminSchema.findById({_id: adminId});
         if(!isPresent) throw new Error('No admin!!!');
+        let isWorkPresent = await worksSchema.findOne({adminId, name});
+        if(isWorkPresent) throw new Error('Work with this name already present!!');
         const workDate = moment().format('MMMM Do YYYY, h:mm:ss a');
         let mongoWork = await worksSchema.create(
             {name, style, adminId, dateOfCreation: workDate});
@@ -31,8 +37,9 @@ module.exports.create = async (req, res) => {
 
 module.exports.deleteWork = async (req, res) => {
   try {
-      let elasticWorks = await client.delete({index: 'works', id: req.params.id});
       let mongoWorks = await worksSchema.findByIdAndDelete({_id: req.params.id});
+      if(!mongoWorks) throw new Error('No Work');
+      let elasticWorks = await client.delete({index: 'works', id: req.params.id});
       res.json({
           success: true,
           elastic: elasticWorks.body,
@@ -50,6 +57,7 @@ module.exports.deleteWork = async (req, res) => {
 module.exports.findWork = async (req, res) => {
   try {
       let work = await worksSchema.findById({_id: req.params.id});
+      if(!work) throw new Error('No Work');
       let lookupData = await Lookup.findOne({key: work.style});
       work.style = lookupData.subject;
       res.json({
@@ -69,6 +77,7 @@ module.exports.findWorksByAdminId = async (req, res) => {
     try{
         let {body} = await client.search({index: 'works',
         body:{query:{match: {adminId: req.params.id}}}});
+        if(!body.length) throw new Error('No Works');
         let works = body.hits.hits;
         let lookup = await Lookup.find({type: 'adminWork'});
         let worksToResponse = works.map(function (work) {
@@ -96,6 +105,7 @@ module.exports.findElasticWork = async (req, res) => {
         //     body: {query:{match: {_id: req.params.id}}}});
         let {body} = await client.get({index: 'works',
             id: req.params.id});
+        if(!body.length) throw new Error('No Works');
         let work = body._source;
         let lookupData = await Lookup.findOne({key: work.style});
         work.style = lookupData.subject;
@@ -139,6 +149,7 @@ module.exports.allWorksFromElastic = async (req, res) => {
 module.exports.findWorkByStyle = async (req, res) => {
     try{
         let elastic = await client.search({index: 'works', body: {query:{match: {style: req.params.style}}}});
+        if(!elastic.body.length) throw new Error('No Works');
         let works = elastic.body.hits.hits;
         let lookup = await Lookup.find({type: 'adminWork'});
         let worksToResponse = works.map(function (work) {
